chore(Security.config): path 이동 #79
